requirements.txt
streamlit>=1.28.0
python-docx==0.8.11
PyPDF2==3.0.1
openai>=1.50.0
python-dotenv==1.0.0 